app/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==1.26.4
PyJWT==2.8.0
scikit-learn==1.4.1.post1
scipy==1.12.0
SQLAlchemy==2.0.29
threadpoolctl==3.4.0
typing_extensions==4.10.0
Werkzeug==3.0.1
